refactor: split MySQL SQL responsibilities
